Added capability to wait for steady state before shutdown
Upon reaching zero connections, we sometimes need to wait
for a few more seconds before shutting the app down. We
are ensuring that all possible residual connections are
accounted for before shutting down.

In the off chance that the app receives traffic, the
steady state timeout will reset.

To disable this feature, set ZEROHALT_DRAIN_STEADY_STATE_WAIT
to zero.

By default, this is set to 5s